made custom shadow component
